# Avoid multiple lookups in `Dictionary::operator[]`