@Bromeon 2025-03-24
> but after looking around a bit I'm fairly certain this isn't causing a behavior change and should work as expected.

Random anecdote: I've used a custom callable (GDExtension) in an untyped dictionary across multiple Godot versions, and counted the number of times the equality operator is invoked:

| Godot version       | Value | Changed in                                                                 |
|---------------------|-------|-----------------------------------------------------------------------|
| 4.1 - 4.3     | 1     |                                                                       |
| 4.4-stable    | 2     |  [#96797](https://github.com/godotengine/godot/pull/96797) |
| `master`      | 1     | this PR                                                   |

So I'm happy for this optimization :slightly_smiling_face: